Move `Image` to `render-core` package.
