+++ src/asset/core/index.js
@@ -1,4 +1,3 @@
-export * from './picture.js'
 export * from './sound.js'
 export * from './asset.js'
 export * from './handle.js'
