+++ src/render-core/assets/image.js
