+++ src/render-core/assets/index.js
@@ -1,3 +1,4 @@
 export * from './material.js'
 export * from './mesh.js'
 export * from './shader.js'
+export * from './image.js'
