+++ src/render-core/plugin.js
@@ -1,7 +1,6 @@
 import { App } from '../app/index.js'
-import { Image } from '../asset/index.js'
 import { Camera, MaterialHandle, MeshHandle } from './components/index.js'
-import { Material, Mesh, Shader } from './assets/index.js'
+import { Material, Mesh, Shader,Image } from './assets/index.js'
 import { ImageParser } from './resources/index.js'
 
 export class RenderCorePlugin {
